+++ Sources/Plasma/PubUtilLib/plCompression/plZlibStream.h
@@ -64,15 +64,15 @@ protected:
     plFileName fFilename;
     const char* fMode;
 
-    int IValidateGzHeader(uint32_t byteCount, const void* buffer);
+    int IValidateGzHeader(size_t byteCount, const void* buffer);
 
 public:
     plZlibStream();
     virtual ~plZlibStream();
 
-    bool     Open(const plFileName& filename, const char* mode) override;
-    bool     Close() override;
-    uint32_t Write(uint32_t byteCount, const void* buffer) override;
+    bool Open(const plFileName& filename, const char* mode) override;
+    bool Close() override;
+    size_t Write(size_t byteCount, const void* buffer) override;
 
     // Since most functions don't check the return value from Write, you can
     // call this after you've passed in all your data to determine if it
@@ -80,12 +80,12 @@ public:
     bool DecompressedOk() { return fDecompressedOk; }
 
     // You can't use these
-    virtual bool     AtEnd();
-    virtual uint32_t Read(uint32_t byteCount, void* buffer);
-    virtual void     Skip(uint32_t deltaByteCount);
-    virtual void     Rewind();
-    virtual void     FastFwd();
-    virtual uint32_t GetEOF();
+    virtual bool AtEnd();
+    virtual size_t Read(size_t byteCount, void* buffer);
+    virtual void Skip(size_t deltaByteCount);
+    virtual void Rewind();
+    virtual void FastFwd();
+    virtual size_t GetEOF();
 };
 
 #endif // plZlibStream_h_inc
